Задания разделены по классам
